+++ sdbitmap/sdbitmap.h
@@ -41,11 +41,11 @@ class sdbitmap {
   
 	uint8_t getBitmap(uint8_t*bmp,    		    // ビットマップデータの切り出し取得(高速版)
 	  uint16_t x, uint16_t y, 
-	  uint8_t w, uint8_t h, uint8_t mode);
+	  uint16_t w, uint16_t h, uint8_t mode);
 
   uint8_t getBitmapEx(uint8_t*bmp,              // ビットマップデータの切り出し取得
     uint16_t x, uint16_t y, 
-    uint8_t w, uint8_t h, uint8_t mode);
+    uint16_t w, uint16_t h, uint8_t mode);
 
 	uint8_t getBitmap(uint8_t*bmp, uint8_t mode);	// ビットマップデータの取得
